perf(index): 稳定 ID 索引——删除按 vector_id 移除，不再全量重建
将索引从「位置即编号」改为「显式稳定 ID」，删除单个文档无需重建整个索引：

- 向量层：FAISS IndexIDMap2（显式 id + remove + 序列化往返）、Milvus add(ids)/remove(ids)、BaseVectorStore 接口扩展

- 稳定 vector_id：sha256(chunk_id) 派生，跨写入稳定可复现；chunk 对象新增 vector_id 字段

- 写入改追加语义：upload 不再覆盖旧索引；metadata.json 按 vector_id 合并；rebuild 先清本地索引再全量写

- MySQL chunks 加 vector_id 列 + 迁移；chunk_repo/Retriever/BM25/ESFulltext/ES 全部改按 vector_id 映射

- 删除：按 vector_id 从 FAISS/Milvus/metadata/ES 移除，无需重建；文档解析加固（document_id/文件名/stem 变体回查）

- 管理台：退出登录清空聊天区等临时状态，避免切换用户残留

实测删除 179s(全量重建)→10.5s(按 ID 移除)；全量测试绿（API74/存储42/单元125，1 个 Milvus 冒烟为环境性 flaky 单独通过）

+++ app/storage/es_client.py
@@ -94,7 +94,7 @@ class ESClient:
                     "document_id": {"type": "keyword"},
                     "strategy": {"type": "keyword"},
                     "chunk_index": {"type": "integer"},
-                    "vector_id": {"type": "integer"},
+                    "vector_id": {"type": "long"},
                     "content": {"type": "text", "analyzer": "ik_max_word"},
                     "start_offset": {"type": "integer"},
                     "end_offset": {"type": "integer"},
